Add files via upload
Four Rooms is a simple game I made when I was 14 in arcade.makecode.com. To open this file, download the png image, go to arcade.makecode.com and click import in the upper right corner. Then select the png download and it will open the game. This game is one of my first creations. I started in the blocks and slowly transitioned to typing in javascript. This game is about a girl that was trapped in a room and must solve multiple puzzles to escape through the main door (the first door the player will see).