Add files via upload
This was one of my final projects for AP Calculus AB. This group project included making a poster on cross sections. We decided to do it after my laptop sticker which was a Donut Juicebox sticker I made. This was a very fun project to do and required some problem solving, especially when it came to creating the function and cross sections for the straw.